% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_current_gw.R
\name{get_current_gw_info}
\alias{get_current_gw_info}
\title{Getting current gameweek information}
\usage{
get_current_gw_info()
}
\description{
This function fetches the current gameweek information
}
\examples{
get_current_gw_info()


}
