% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_current_gw.R
\name{get_current_gw_number}
\alias{get_current_gw_number}
\title{Getting current gameweek number}
\usage{
get_current_gw_number()
}
\description{
This function fetches the current gameweek number
}
\examples{

get_current_gw_number()

}
