% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_ep_for_entrant.R
\name{get_ep_for_entrant}
\alias{get_ep_for_entrant}
\title{Getting Expected Points for a entrant in a given gameweek}
\usage{
get_ep_for_entrant(entrant_number, gameweek)
}
\arguments{
\item{entrant_number}{the entrant number you want a table for}

\item{gameweek}{the gameweek for expected point}
}
\description{
This function returns expected points for an entrant
in a given gameweek.
}
\examples{
# Getting the expected point for my team in GW1
get_ep_for_league(9680, 1)

}
