% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_ep_for_league.R
\name{get_ep_for_league}
\alias{get_ep_for_league}
\title{Getting Expected Points for a league in a given gameweek}
\usage{
get_ep_for_league(league_number, gameweek)
}
\arguments{
\item{league_number}{the minileague number you want a table of}

\item{gameweek}{the gameweek for expected point}
}
\description{
This function returns an expected points table for a minileague
in a given gameweek.
}
\examples{
# Getting the expected point for my league in GW1
get_ep_for_league(570437, 1)

}
