% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_ep_not_owned.R
\name{get_ep_not_owned}
\alias{get_ep_not_owned}
\title{Getting Expected Points for players not owned}
\usage{
get_ep_not_owned(entrant_number, gameweek)
}
\arguments{
\item{entrant_number}{the entrant number you want a table for}

\item{gameweek}{the gameweek for expected point}
}
\description{
This function players with high expected points
not owned by the entrant in a given gameweek.
}
\examples{
# Getting the expected point for players not owned by entrant
get_ep_not_owned(9680, 1)

}
