% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_fdr_by_gameweek.R
\name{get_fdr_by_gameweek}
\alias{get_fdr_by_gameweek}
\title{Getting fdr table by gameweek}
\usage{
get_fdr_by_gameweek()
}
\description{
This function combines the fixtures and fdr ratings to get
a table by gamweek.
}
\examples{
get_fdr_by_gameweek()


}
