% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_fdr_by_gameweek.R
\name{get_fdr_for_selected_gameweek}
\alias{get_fdr_for_selected_gameweek}
\title{Getting fdr table by specific gameweeks}
\usage{
get_fdr_for_selected_gameweek(input_gw, input_type = "overall")
}
\arguments{
\item{input_gw}{the gameweeks you want the fdr table for}

\item{input_type}{one of "overall", "attack", "defence". Defaults to "overall"}
}
\description{
This function combines the fixtures and fdr ratings to get
a table by gameweek, and then select specific columns for gameweeks specified.
If you want the table to be update according to user input, these should be reactive
}
\examples{
gw_columns <- c(30:35)
input_type <- "attack"

get_fdr_for_selected_gameweek(gw_columns, input_type)


}
