% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_league_stats.R
\name{get_league_bench_points}
\alias{get_league_bench_points}
\title{Getting a table of bench points for a minileague}
\usage{
get_league_bench_points(league_number)
}
\arguments{
\item{league_number}{the league number you want a table for}
}
\description{
This function returns a table of
bench points for a minileague. Automatically return a
table for cumulative amount.
}
\examples{
# Getting the minileague information
get_league_bench_points(570437)

}
