% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_league_stats.R
\name{get_league_general_info}
\alias{get_league_general_info}
\title{Getting a table of general information for a minileague's current gw}
\usage{
get_league_general_info(league_number)
}
\arguments{
\item{league_number}{the league number you want a table for}
}
\description{
This function returns a table of
general information for a minileague. Automatically return a
table for the current gameweek.
}
\examples{
# Getting the minileague information
get_league_general_info(570437)

}
