% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_historic_league_standing.R
\name{get_league_historic_standings}
\alias{get_league_historic_standings}
\title{Getting historic league standing}
\usage{
get_league_historic_standings(league_number)
}
\arguments{
\item{league_number}{the minileague number you want a table of}
}
\description{
This function returns a mini league's standings over time
}
\examples{
# Getting standings for my league across weeks
get_league_historic_standings(570437)

}
