% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_league_stats.R
\name{get_league_hits_taken}
\alias{get_league_hits_taken}
\title{Getting a table of hits taken for a minileague}
\usage{
get_league_hits_taken(league_number)
}
\arguments{
\item{league_number}{the league number you want a table for}
}
\description{
This function returns a table of
points deducted for transfers for a minileague. Automatically return a
cumulative table.
}
\examples{
# Getting the minileague information
get_league_team_value(570437)

}
