% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_historic_league_standing.R
\name{get_league_name}
\alias{get_league_name}
\title{Getting league name}
\usage{
get_league_name(league_number)
}
\arguments{
\item{league_number}{the minileague number you want the name of}
}
\description{
This function returns the league name in character format
}
\examples{
# Getting the league name for my league
get_league_name(570437)

}
