% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_league_stats.R
\name{get_league_team_value}
\alias{get_league_team_value}
\title{Getting a table of team value for a minileague}
\usage{
get_league_team_value(league_number)
}
\arguments{
\item{league_number}{the league number you want a table for}
}
\description{
This function returns a table of
team value for a minileague. Automatically return a
table for current gameweek.
}
\examples{
# Getting the minileague information
get_league_team_value(570437)

}
