% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_current_gw.R
\name{get_next_gw_info}
\alias{get_next_gw_info}
\title{Getting next gameweek information}
\usage{
get_next_gw_info()
}
\description{
This function fetches the next gameweek information
}
\examples{
get_next_gw_info()


}
