% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_current_gw.R
\name{get_next_gw_number}
\alias{get_next_gw_number}
\title{Getting next gameweek number}
\usage{
get_next_gw_number()
}
\description{
This function fetches the next gameweek number
}
\examples{

get_next_gw_number()

}
