% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_players_owned.R
\name{get_players_owned}
\alias{get_players_owned}
\title{Get players owned}
\usage{
get_players_owned(entrant_number, gameweeks)
}
\arguments{
\item{entrant_number}{the entrant number you want a table for}

\item{gameweeks}{gameweeks data should be fetched for}
}
\description{
Fetches the players the entrant owned for given gameweeks. This can then
be fed into a plot of players owned, or starting 11 throughout the weeks.
}
\examples{
# Getting the expected point for my team in GW1
get_players_owned(9680, 1)

# Getting the expected points for my team GW 1-16
get_players_owned(9680, c(1:16))

}
