% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reactable_helpers.R
\name{get_rct_columns}
\alias{get_rct_columns}
\title{Getting column list for reactable}
\usage{
get_rct_columns(fdr_selected_gameweek, input_gw)
}
\arguments{
\item{fdr_selected_gameweek}{the fdr_selected_gameweek() functions}

\item{input_gw}{the gameweek provided by user selection}
}
\description{
Returning a list of column definitions for a reactable table.
}
