% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reactable_helpers.R
\name{get_rct_groups}
\alias{get_rct_groups}
\title{Getting column groups for reactable}
\usage{
get_rct_groups(input_gw)
}
\arguments{
\item{input_gw}{the gameweek provided by user selection}
}
\description{
Returning a list of column groups for a reactable table.
}
