% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_conditional_themes.R
\name{gt_table_theme}
\alias{gt_table_theme}
\title{Getting a gt table theme according to light/darkmode}
\usage{
gt_table_theme(data, current_theme)
}
\arguments{
\item{data}{the gt table object}

\item{current_theme}{the theme is format "dark" or "light"}
}
\description{
This function returns a gt table theme depending on whether
the app is in light or dark more
}
