% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_themes.R
\docType{data}
\name{light}
\alias{light}
\title{bs_themes for light mode}
\format{
An object of class \code{bs_version_5} (inherits from \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
light
}
\description{
bs_themes for light mode
}
\keyword{datasets}
