% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots_league.R
\name{plot_league_points}
\alias{plot_league_points}
\title{Plotting league points}
\usage{
plot_league_points(league_number)
}
\arguments{
\item{league_number}{the minileague number you want a table of}
}
\description{
This function plots mini league's points over time
}
\examples{
# Getting standings for my league across weeks
plot_league_points(570437)

}
