% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots_league.R
\name{plot_league_standings}
\alias{plot_league_standings}
\title{Plotting league standing}
\usage{
plot_league_standings(league_number)
}
\arguments{
\item{league_number}{the minileague number you want a table of}
}
\description{
This function plots mini league's standings over time
}
\examples{
# Getting standings for my league across weeks
plot_league_standings(570437)

}
