% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_players_owned.R
\name{plot_players_owned}
\alias{plot_players_owned}
\title{Plot players owned throughout gameweeks}
\usage{
plot_players_owned(entrant_number, gameweeks)
}
\arguments{
\item{entrant_number}{the team number for your team}

\item{gameweeks}{the gameweeks you want to plot for}
}
\description{
Plots all of the players an entrant has owned for each gameweek
}
\examples{
# Getting players owned for my team across gameweeks
plot_players_owned(9680, c(1:16))

}
