% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_players_owned.R
\name{plot_starting_eleven}
\alias{plot_starting_eleven}
\title{Plot the starting eleven throughout gameweeks}
\usage{
plot_starting_eleven(entrant_number, gameweeks)
}
\arguments{
\item{entrant_number}{the team number for your team}

\item{gameweeks}{the gameweeks you want to plot for}
}
\description{
Plots all of the players by position for each gameweek
}
\examples{
# Getting players owned for my team across gameweeks
plot_starting_eleven(9680, c(1:16))

}
