% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_conditional_themes.R
\name{reactable_table_theme}
\alias{reactable_table_theme}
\title{Getting a reactable table theme according to light/darkmode}
\usage{
reactable_table_theme(current_theme)
}
\arguments{
\item{current_theme}{the theme is format "dark" or "light"}
}
\description{
This function returns a table of
general information for a minileague. Automatically return a
table for the current gameweek.
}
