% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_themes.R
\name{gt_table_theme}
\alias{gt_table_theme}
\title{Getting a gt table theme according to light/darkmode}
\usage{
gt_table_theme(data, current_theme)
}
\arguments{
\item{data}{the gt table object}

\item{current_theme}{the theme is format "dark" or "light"}
}
\description{
This function returns a table of
general information for a minileague. Automatically return a
table for the current gameweek.
}
